+++ types-create/types/augmentRegistry.ts
@@ -3,12 +3,12 @@
 
 // import type lookup before we augment - in some environments
 // this is required to allow for ambient/previous definitions
-import 'https://deno.land/x/polkadot@0.2.20/types-codec/types/registry.ts';
+import 'https://deno.land/x/polkadot/types-codec/types/registry.ts';
 
-import type { Codec, CodecClass, ICompact, INumber } from 'https://deno.land/x/polkadot@0.2.20/types-codec/types/index.ts';
-import type { ILookup, TypeDef } from 'https://deno.land/x/polkadot@0.2.20/types-create/types/index.ts';
+import type { Codec, CodecClass, ICompact, INumber } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
+import type { ILookup, TypeDef } from 'https://deno.land/x/polkadot/types-create/types/index.ts';
 
-declare module 'https://deno.land/x/polkadot@0.2.20/types-codec/types/registry.ts' {
+declare module 'https://deno.land/x/polkadot/types-codec/types/registry.ts' {
   interface Registry {
     readonly lookup: ILookup;