+++ types/interfaces/offchain/types.ts
@@ -1,7 +1,7 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { Enum } from 'https://deno.land/x/polkadot@0.2.20/types-codec/mod.ts';
+import type { Enum } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
 
 /** @name StorageKind */
 export interface StorageKind extends Enum {