+++ types/interfaces/scaleInfo/definitions.ts
@@ -3,7 +3,7 @@
 
 import type { Definitions } from '../../types/index.ts';
 
-import { objectSpread } from 'https://deno.land/x/polkadot@0.2.20/util/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { v0 } from './v0.ts';
 import { v1 } from './v1.ts';